Refactor: 채팅방 리스트 관련 파일을 RecentChat 폴더 내부로 이동
